Coded the basics of each necessary files so they work and interact. Also created a Docs folder (by accident) and added the WindowBuilder tab.
The Controller and Runner contains code so that they can actually run the program with the Controller containing info that the Frame will use. The Frame takes some of that and then create the dimensions and information for the Panel (which can be changed if needed) while the Panel has the basic layout of what is needed. All also contains the imports that they will need at the moment.