Created a new file to be able to read the file, added said file to the project, created a PopupDisplay, and added various imports.
The file reader still needs some tweaks to see if it works well with the file. And the popupDisplay may not be necessary at all. The controller will hopefully send the information and eventually it can all be processed in the Tarot.java file to do things with the Panel through action listeners.