Removed the readCardType method and started coding for Tarot.java.
readCardType is not needed for the way I am going to do it. Tarot.java will be taking the "contents" of the file reader and splitting the information by each new line. Then it will be added to an array to work with. The TarotCardMeanings.txt has been moved to the root folder.