Added a method called ClearOldCards() and uncommented the removers as well as an @Override fo rthe drawCardsButton.
The issue with the button not working was due to the fact that CardInfo did not clear out previous entries. The ClearOldCards() method clears the deck and the CardInfo arrays before the new set of cards is created and 3 new ones are selected. Making sure the random number is not bigger or equal to the deck size made crashes happen less often. But it can still occur. Also sometimes the image file cannot be found sometimes. But overall the program functions as intended for the most part.