Removed the MainMenuPanel & created a CardLayout class that'll hold multiple panels.
The TarotCardLayout will create a CardLayout class that is able to hold panels within it. This will allow the use of a menu screen and allow panels to go from one to another with ease. Although the new issue is how to get seperate panel classes like TarotPanel to work with it without having to copy everything over.