Got the TarotCardLayout to display the TarotPanel by clicking a button.
GOT IT WORKING!
The trick was that there was no necessary inheritance or imports. You simply need to create an object of the class of the panel you want. And then add it to the cardLayout like you would another panel and then do the same button Action Listeners as needed.