Added & Fixed descriptions for TarotCrossSpread and started to attempt to code the menu button.
The menu button still doesn't work like it should. It creates a second window that goes back to the cardLayout main screen but keeps the old one. As for the text issue with the card descriptions, I just needed to change the size of the text box and make it wrap.